refactor: 캐릭터 등급 관련 enum 변경
- enum관리를 위해 common패키지로 위치 변경
- 등급 산출 메서드 수정
